requirements.txt
dhooks~=1.1.3
discord.py~=2.3.2
discord-pretty-help~=2.0.1
asyncpraw~=7.7.1
aiohttp~=3.9.1
aiofiles~=23.2.0
pandas~=2.1.4
snscrape~=0.7.0
vadersentiment~=3.3.2
tradingview-ta~=3.3.0
asyncio~=3.4.3
